requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiobotocore[boto3]==1.3.0
    # via
    #   dvc
    #   s3fs
aiohttp==3.7.4.post0
    # via aiobotocore
aioitertools==0.7.1
    # via aiobotocore
aniso8601==7.0.0
    # via graphene
anyascii==0.2.0
    # via wagtail
appdirs==1.4.4
    # via
    #   dvc
    #   dvc-pandas
asgiref==3.3.4
    # via django
async-timeout==3.0.1
    # via aiohttp
atpublic==2.3
    # via flufl.lock
attrs==21.2.0
    # via
    #   aiohttp
    #   pytest
backcall==0.2.0
    # via ipython
beautifulsoup4==4.9.3
    # via wagtail
boto3==1.17.49
    # via aiobotocore
botocore==1.20.49
    # via
    #   aiobotocore
    #   boto3
    #   s3transfer
certifi==2021.5.30
    # via
    #   dulwich
    #   requests
    #   sentry-sdk
cffi==1.14.5
    # via pygit2
chardet==4.0.0
    # via
    #   aiohttp
    #   requests
colorama==0.4.4
    # via
    #   dvc
    #   rich
colored==1.4.2
    # via -r requirements.in
commonmark==0.9.1
    # via rich
configobj==5.0.6
    # via dvc
decorator==4.4.2
    # via
    #   ipython
    #   jsonpath-ng
    #   networkx
dictdiffer==0.8.1
    # via dvc
diskcache==5.2.1
    # via dvc
distro==1.5.0
    # via dvc
django-cors-headers==3.7.0
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==2.4.0
    # via wagtail
django-modelcluster==5.1
    # via wagtail
django-stubs-ext==0.2.0
    # via django-stubs
django-stubs==1.8.0
    # via -r requirements.in
django-taggit==1.4.0
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
django==3.2.4
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-stubs
    #   django-stubs-ext
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   graphene-django
    #   wagtail
    #   wagtailfontawesome
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   wagtail
dpath==2.0.1
    # via dvc
draftjs-exporter==2.1.7
    # via wagtail
dulwich==0.20.23
    # via dvc
dvc-pandas==0.0.6
    # via -r requirements.in
dvc[s3]==2.3.0
    # via dvc-pandas
et-xmlfile==1.1.0
    # via openpyxl
factory-boy==3.2.0
    # via
    #   -r requirements.in
    #   pytest-factoryboy
faker==8.7.0
    # via factory-boy
flatten-dict==0.4.0
    # via dvc
flufl.lock==3.2
    # via dvc
fsspec==2021.05.0
    # via
    #   dvc
    #   s3fs
ftfy==6.0.3
    # via python-benedict
funcy==1.16
    # via dvc
future==0.18.2
    # via grandalf
gitdb==4.0.7
    # via gitpython
gitpython==3.1.17
    # via
    #   dvc
    #   dvc-pandas
grandalf==0.6
    # via dvc
graphene-django-optimizer==0.8.0
    # via -r requirements.in
graphene-django==2.15.0
    # via -r requirements.in
graphene-stubs==0.15
    # via -r requirements.in
graphene==2.1.8
    # via graphene-django
graphql-core==2.3.2
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==2.0.1
    # via graphene
html5lib==1.1
    # via wagtail
idna==2.10
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==1.1.1
    # via pytest
ipython-genutils==0.2.0
    # via traitlets
ipython==7.24.1
    # via -r requirements.in
jedi==0.18.0
    # via ipython
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonpath-ng==1.5.2
    # via dvc
l18n==2020.6.1
    # via wagtail
mailchecker==4.0.8
    # via python-benedict
matplotlib-inline==0.1.2
    # via ipython
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via mypy
mypy==0.902
    # via
    #   django-stubs
    #   graphene-stubs
nanotime==0.5.2
    # via dvc
networkx==2.5.1
    # via dvc
numpy==1.20.3
    # via
    #   pandas
    #   pyarrow
    #   scipy
openpyxl==3.0.7
    # via tablib
orjson==3.5.3
    # via -r requirements.in
packaging==20.9
    # via
    #   dvc
    #   pint
    #   pytest
pandas==1.2.4
    # via
    #   -r requirements.in
    #   dvc-pandas
    #   pint-pandas
parso==0.8.2
    # via jedi
pathspec==0.8.1
    # via dvc
pexpect==4.8.0
    # via ipython
phonenumbers==8.12.24
    # via python-benedict
pickleshare==0.7.5
    # via ipython
pillow==8.2.0
    # via wagtail
pint-pandas==0.2
    # via
    #   -r requirements.in
    #   dvc-pandas
pint==0.17
    # via
    #   -r requirements.in
    #   pint-pandas
pluggy==0.13.1
    # via pytest
ply==3.11
    # via
    #   dvc
    #   jsonpath-ng
promise==2.3
    # via
    #   graphene-django
    #   graphql-core
    #   graphql-relay
prompt-toolkit==3.0.18
    # via ipython
psutil==5.8.0
    # via dvc
psycopg2-binary==2.8.6
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pyarrow==4.0.1
    # via dvc-pandas
pyasn1==0.4.8
    # via dvc
pycparser==2.20
    # via cffi
pydot==1.4.2
    # via dvc
pygit2==1.6.0
    # via dvc
pygments==2.9.0
    # via
    #   ipython
    #   rich
pygtrie==2.3.2
    # via dvc
pyparsing==2.4.7
    # via
    #   dvc
    #   grandalf
    #   packaging
    #   pydot
pytest-django==4.4.0
    # via -r requirements.in
pytest-factoryboy==2.1.0
    # via -r requirements.in
pytest==6.2.4
    # via
    #   pytest-django
    #   pytest-factoryboy
python-benedict==0.24.0
    # via dvc
python-dateutil==2.8.1
    # via
    #   botocore
    #   faker
    #   pandas
    #   python-benedict
python-dotenv==0.17.1
    # via -r requirements.in
python-fsutil==0.5.0
    # via python-benedict
python-slugify==5.0.2
    # via python-benedict
python-snappy==0.6.0
    # via dvc-pandas
pytz==2021.1
    # via
    #   django
    #   django-modelcluster
    #   l18n
    #   pandas
pyyaml==5.4.1
    # via python-benedict
redis==3.5.3
    # via -r requirements.in
requests==2.25.1
    # via
    #   dvc
    #   python-benedict
    #   python-fsutil
    #   wagtail
rich==10.3.0
    # via dvc
ruamel.yaml.clib==0.2.2
    # via ruamel.yaml
ruamel.yaml==0.17.9
    # via
    #   dvc
    #   dvc-pandas
rx==1.6.1
    # via graphql-core
s3fs==2021.5.0
    # via dvc
s3transfer==0.3.7
    # via boto3
scipy==1.6.3
    # via -r requirements.in
sentry-sdk==1.1.0
    # via -r requirements.in
shortuuid==1.0.1
    # via dvc
shtab==1.3.6
    # via dvc
singledispatch==3.6.2
    # via graphene-django
six==1.16.0
    # via
    #   configobj
    #   flatten-dict
    #   graphene
    #   graphene-django
    #   graphql-core
    #   graphql-relay
    #   html5lib
    #   jsonpath-ng
    #   l18n
    #   promise
    #   python-benedict
    #   python-dateutil
    #   singledispatch
smmap==4.0.0
    # via gitdb
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.1
    # via django
tablib[xls,xlsx]==3.0.0
    # via wagtail
tabulate==0.8.9
    # via dvc
telepath==0.2
    # via wagtail
text-unidecode==1.3
    # via
    #   faker
    #   graphene-django
    #   python-slugify
toml==0.10.2
    # via
    #   dvc
    #   mypy
    #   pytest
    #   python-benedict
tqdm==4.61.0
    # via dvc
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
types-orjson==0.1.1
    # via -r requirements.in
typing-extensions==3.10.0.0
    # via
    #   aiohttp
    #   aioitertools
    #   django-stubs
    #   dvc
    #   graphene-stubs
    #   mypy
urllib3==1.26.5
    # via
    #   botocore
    #   dulwich
    #   requests
    #   sentry-sdk
voluptuous==0.12.1
    # via dvc
wagtail==2.13.1
    # via
    #   -r requirements.in
    #   wagtailfontawesome
wagtailfontawesome==1.2.1
    # via -r requirements.in
wcwidth==0.2.5
    # via
    #   ftfy
    #   prompt-toolkit
webencodings==0.5.1
    # via html5lib
willow==1.4
    # via wagtail
wrapt==1.12.1
    # via aiobotocore
xlrd==2.0.1
    # via tablib
xlsxwriter==1.4.3
    # via wagtail
xlwt==1.3.0
    # via tablib
xmltodict==0.12.0
    # via python-benedict
yarl==1.6.3
    # via aiohttp
zc.lockfile==2.0
    # via dvc

# The following packages are considered to be unsafe in a requirements file:
# setuptools
